Moved facilities to its own table (HotelFacilities)
The facilities are stored in a SaaS model, where each facility has a hotel_id
along with the name as unique primary & range keys. Now search can be done based
on hotel_id and f_name. Based on this, the Facilities.js db class is used to fetch the data

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>